Entity collisions, gravity, & velocity. - Refactor Collision, fix deltatime bugs. - Apply gravity to applicable entities. - Add callbacks to ItemDef, unused so far. - Reorganize and document LocalPlayer class. - Give bees, rabbits, and test entities AI that is *so cute guys seriously look at it*. - Plains biomes now have flowers.

+++ src/util/Target.h
@@ -12,7 +12,7 @@ class Target {
 public:
 	
 	/** A struct containing data for a targeted entity. */
-	struct EntityData { unsigned int id; };
+	struct EntityData { long long id; };
 	/** A struct containing data for a targeted block & face. */
 	struct BlockData { glm::ivec3 pos; EVec face; };
 	/** A struct containing no data for targets pointing to nothing. */
